Followed the A.04.1 instructions
Added a vertical layout to organise the UI text and the button.
Added a "image" folder" in the "Assets" folder.
Added an image of a Lamborghini huracan in the 'image' folder.
Added the text and their definition in the script.
Added a interactive button.
Added a function "OnClickDecreaseFuel" in the script to manage the button.
Added a function "Awake" to manage the text UI reference before the start function.
Commented out all the print functions.
Resolved the conflicts.